Rename part 0: /order to /register path ops
"Order" has become a pretty overloaded term in this project. In the
following commits, I introduce a few renames in hope to reduce various
terminology confusions.

This first commit renames the `/order` path operations to `/register`. I
chose the term "register" because the page acts kind of like a cash
register: the page presents various buttons to properly handle incoming
order requests. This clears up the distinction between the register as a
machine and order sessions as intermediate states.